requirements.txt
adal==1.2.7
antlr4-python3-runtime==4.9.3
applicationinsights==0.11.10
argcomplete==2.0.0
azure-appconfiguration==1.1.1
azure-batch==13.0.0
azure-cli==2.45.0
azure-cli-core==2.45.0
azure-cli-telemetry==1.0.8
azure-common==1.1.28
azure-core==1.26.3
azure-cosmos==3.2.0
azure-data-tables==12.4.0
azure-datalake-store==0.0.52
azure-functions==1.13.2
azure-graphrbac==0.60.0
azure-identity==1.12.0
azure-keyvault==1.1.0
azure-keyvault-administration==4.0.0b3
azure-keyvault-keys==4.8.0b2
azure-loganalytics==0.1.1
azure-mgmt-advisor==9.0.0
azure-mgmt-apimanagement==3.0.0
azure-mgmt-appconfiguration==2.2.0
azure-mgmt-applicationinsights==1.0.0
azure-mgmt-authorization==0.61.0
azure-mgmt-batch==17.0.0
azure-mgmt-batchai==7.0.0b1
azure-mgmt-billing==6.0.0
azure-mgmt-botservice==2.0.0
azure-mgmt-cdn==12.0.0
azure-mgmt-cognitiveservices==13.3.0
azure-mgmt-compute==29.1.0
azure-mgmt-consumption==2.0.0
azure-mgmt-containerinstance==9.1.0
azure-mgmt-containerregistry==10.0.0
azure-mgmt-containerservice==21.1.0
azure-mgmt-core==1.3.2
azure-mgmt-cosmosdb==8.0.0
azure-mgmt-databoxedge==1.0.0
azure-mgmt-datalake-analytics==0.2.1
azure-mgmt-datalake-nspkg==3.0.1
azure-mgmt-datalake-store==0.5.0
azure-mgmt-datamigration==10.0.0
azure-mgmt-deploymentmanager==0.2.0
azure-mgmt-devtestlabs==4.0.0
azure-mgmt-dns==8.0.0
azure-mgmt-eventgrid==10.2.0b2
azure-mgmt-eventhub==10.1.0
azure-mgmt-extendedlocation==1.0.0b2
azure-mgmt-hdinsight==9.0.0
azure-mgmt-imagebuilder==1.1.0
azure-mgmt-iotcentral==10.0.0b2
azure-mgmt-iothub==2.3.0
azure-mgmt-iothubprovisioningservices==1.1.0
azure-mgmt-keyvault==10.1.0
azure-mgmt-kubernetesconfiguration==2.0.0
azure-mgmt-kusto==0.3.0
azure-mgmt-loganalytics==13.0.0b4
azure-mgmt-managedservices==1.0.0
azure-mgmt-managementgroups==1.0.0
azure-mgmt-maps==2.0.0
azure-mgmt-marketplaceordering==1.1.0
azure-mgmt-media==9.0.0
azure-mgmt-monitor==5.0.1
azure-mgmt-msi==6.1.0
azure-mgmt-netapp==9.0.1
azure-mgmt-network==21.0.1
azure-mgmt-nspkg==3.0.2
azure-mgmt-policyinsights==1.1.0b4
azure-mgmt-privatedns==1.0.0
azure-mgmt-rdbms==10.2.0b6
azure-mgmt-recoveryservices==2.2.0
azure-mgmt-recoveryservicesbackup==5.1.0
azure-mgmt-redhatopenshift==1.2.0
azure-mgmt-redis==14.1.0
azure-mgmt-relay==0.1.0
azure-mgmt-resource==21.1.0b1
azure-mgmt-search==8.0.0
azure-mgmt-security==3.0.0
azure-mgmt-servicebus==8.1.0
azure-mgmt-servicefabric==1.0.0
azure-mgmt-servicefabricmanagedclusters==1.0.0
azure-mgmt-servicelinker==1.2.0b1
azure-mgmt-signalr==1.1.0
azure-mgmt-sql==4.0.0b7
azure-mgmt-sqlvirtualmachine==1.0.0b4
azure-mgmt-storage==21.0.0
azure-mgmt-synapse==2.1.0b5
azure-mgmt-trafficmanager==1.0.0
azure-mgmt-web==7.0.0
azure-multiapi-storage==1.0.0
azure-nspkg==3.0.2
azure-storage-common==1.4.2
azure-synapse-accesscontrol==0.5.0
azure-synapse-artifacts==0.14.0
azure-synapse-managedprivateendpoints==0.3.0
azure-synapse-spark==0.2.0
bcrypt==4.0.1
black==23.1.0
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
chardet==3.0.4
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
cryptography==39.0.1
Deprecated==1.2.13
distro==1.8.0
fabric==2.7.1
google-api-core==2.11.0
google-auth==2.16.1
googleapis-common-protos==1.58.0
humanfriendly==10.0
idna==3.4
invoke==1.7.3
isodate==0.6.1
javaproperties==0.5.2
jmespath==1.0.1
jsondiff==2.0.0
knack==0.10.1
msal==1.20.0
msal-extensions==1.0.0
msrest==0.7.1
msrestazure==0.6.4
mypy-extensions==1.0.0
oauthlib==3.2.2
opencensus==0.11.1
opencensus-context==0.1.3
opencensus-ext-azure==1.1.8
packaging==21.3
paramiko==2.12.0
pathlib2==2.3.7.post1
pathspec==0.11.0
pkginfo==1.9.6
platformdirs==2.6.2
portalocker==2.7.0
protobuf==4.22.0
psutil==5.9.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyGithub==1.58.0
Pygments==2.14.0
PyJWT==2.6.0
PyMySQL==1.0.2
PyNaCl==1.5.0
pyOpenSSL==23.0.0
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
PyYAML==6.0
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
scp==0.13.6
semver==2.13.0
six==1.16.0
sshtunnel==0.1.5
tabulate==0.9.0
tomli==2.0.1
typing_extensions==4.5.0
urllib3==1.26.14
urllib3-secure-extra==0.1.0
websocket-client==1.3.3
wrapt==1.15.0
xmltodict==0.13.0
